refactor: split read_model session.rs into focused modules
session.rs (1,507 lines, no Session type) mixed four concerns. Move them
into capabilities.rs (adapter/query capability descriptors), mutation.rs
(staged mutations + row/key validation helpers), plan.rs (write plan +
detached builder), load.rs (load request/graph/builder), and workspace.rs
(store-bound tracked workspace). Pure move; public paths preserved via
re-exports in read_model/mod.rs.

Implements [[tasks/read-model-perf-structure]]
